Новые изменения

+++ Пр12/Cart.cs
@@ -23,7 +23,7 @@ namespace Пр12
         public int ID { get; set; }
         public int UserID { get; set; }
         public decimal TotalAmount { get; set; }
-        public Nullable<decimal> TotalQuantity { get; set; }
+        public Nullable<int> TotalQuantity { get; set; }
     
         public virtual User User { get; set; }
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
